backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
boto3==1.34.131
botocore==1.34.131
python-dotenv==1.0.0
python-multipart==0.0.6
kubernetes==29.0.0
opensearch-py==2.4.2
requests-aws4auth==1.2.3
pinecone-client==3.0.2
langchain==0.1.4
langchain-community==0.0.16
sentence-transformers==2.3.1
numpy==1.26.3
pandas==2.2.0
requests==2.31.0
httpx==0.26.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.1
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
structlog==24.1.0
